SQLite database for managing Spawnables (Inventory) and world saves. (#168)
* Added the godot-SQLite addon.

Added a basic SQLite database for the application.
Assets are added to the database.
Assets now properly hash their file contents before going onto the disk.
Minor benchmarking notes.
Note: The godot-SQLite addon is not in the repository as I look for a
good way to include this addon without actually commiting heavy binaries
to the repository. (Version 4.8 of the addon was used at this time.)

* Save spawnables.

Adds basic support for adding spawnables to the database.

* Spawnable dashboard page uses database to get filelist.

Added thumbnail column to the spawnable database.
Set the creation date when saving a spawnable.

* Fixed loading spawnables.

* Basic world loading from spawnable menu.

When double clicking on the world in the spawnables menu, you are now
prompted to create a new world with the selected world as the base.

* Better world loading handling.

When loading a world, the create new world dialog is reset.
When closing the dialog, it is reset.
Moved dialog code to dedicated script file.
Added "CUSTOM" base world type. It's only to be used when loading a
world from disk.

* Open New World dialog in center of scren.

When the dialog is to be shown, it is moved to the center of the screen.

* Delete spawnables.

This fixes deleting spawnbles, and will now also delete the spawnables
from the database.

* Spawnable asset relationship.

Adds a table describing the relationship between assets and spawnables.
If a spawnable is deleted, all assets related to that spawnable (if they
are not being used by another spawnable) are deleted.

* Added setup instructions and install script for linux.

Since the modules are not shipped with the project, provide an install
script so that a developer can quickly get the addons required.
Also provide instructions on how to download the files manually.

+++ src/userinterface/dash/scripts/pages/spawnables.gd
@@ -6,7 +6,6 @@
 # License: MIT License
 # Authors: Armored Dragon
 # --- License
-
 extends "res://userinterface/dash/scripts/pages/left_nav_container.gd"
 
 const BASE_DIR = "user://inventory/"
@@ -20,12 +19,12 @@ var _selected_folder: Node
 @onready var _path_container = $"HBox/Right/Local/Path"
 @onready var _folder_container = $"HBox/Right/Local/Folders"
 @onready var _file_container = $"HBox/Right/Local/Files"
-
 @onready var _create_folder_btn = $"HBox/Right/Local/Actions/NewFolder/Button"
 @onready var _delete_selected_btn = $"HBox/Right/Local/Actions/DeleteSelected/Button"
-
 @onready var _folder_creation_dialog = $"FolderCreation"
 @onready var _deletion_dialog = $"DeleteConfirmation"
+@onready var dashboard = get_tree().current_scene.get_node("Dashboard")
+
 
 func _ready():
 	super._ready()
@@ -41,11 +40,13 @@ func _ready():
 	_build_view()
 	return
 
+
 func _handle_page_opened(page_name) -> void:
 	if page_name != "Spawnables":
 		return
 
 	return
+
 
 func _build_view() -> void:
 	var _filelist = FileManager.get_inv_filelist()
@@ -55,13 +56,24 @@ func _build_view() -> void:
 	await _clear_view()
 
 	for file in _files:
-		var _file_name = file.get_basename()
-		var _listing = _create_button(_file_name, "", _load_file.bind(file))
+		var _file_pretty_name = file.name
+		var _file_name: String = file.hash
+		var _listing
+		if file.type == Enum.SpawnableType.WORLD:
+			_listing = _create_button(_file_pretty_name, "", _create_world.bind(_file_name + ".tscn"))
+		else:
+			_listing = _create_button(_file_pretty_name, "", _load_file.bind(_file_name + ".tscn"))
+
+		_listing.set_meta("type", "spawnable")
+		_listing.set_meta("file_name", _file_name)
 		_file_container.add_child(_listing)
 		continue
 
 	for folder in _dirs:
 		var _listing = _create_button(folder, "Folder", _dir_deeper.bind(folder))
+
+		_listing.set_meta("type", "folder")
+
 		_folder_container.add_child(_listing)
 		continue
 
@@ -88,20 +100,37 @@ func _build_view() -> void:
 		continue
 	return
 
+
 func _dir_deeper(path: String) -> void:
 	FileManager.move_inv_deeper(path)
 	_build_view()
 	return
+
 
 func _dir_relocate(index: int) -> void:
 	FileManager.move_inv_relocate(index + 1)
 	_build_view()
 	return
 
+
 func _load_file(file_name: String) -> void:
-	var _path = FileManager._current_path() + "/" + file_name
+	var _path = FileManager._current_path() + file_name
 	_app_spawnable_file_handling_m.load_spawnable(_path)
 	return
+
+
+func _create_world(file_name: String) -> void:
+	var _dialog: Control = dashboard.get_node("NewWorld")
+	var _world_load_dir_val: Control = _dialog.get_node("%WorldLoadDirVal")
+	var _base_val: Control = _dialog.get_node("%BaseVal")
+	var _file_path = FileManager._current_path() + file_name
+
+	_world_load_dir_val.text = _file_path
+	_base_val.selected = Enum.BaseLevel.CUSTOM
+
+	_dialog.show_window()
+	return
+
 
 func _create_button(btn_name: String, icon: String = "", double_click = null, min_size: Vector2 = Vector2(350, 50)) -> Node:
 	var _listing = _template_button.instantiate()
@@ -119,6 +148,7 @@ func _create_button(btn_name: String, icon: String = "", double_click = null, mi
 	_action_buttons.append(_listing)
 	return _listing
 
+
 func _button_selected(selected: Node) -> void:
 	var _button_name = selected.get_meta("label")
 
@@ -129,6 +159,7 @@ func _button_selected(selected: Node) -> void:
 
 	_selected_folder = selected
 	return
+
 
 func _clear_view() -> void:
 	_action_buttons = []
@@ -147,9 +178,11 @@ func _clear_view() -> void:
 
 	return
 
+
 func _open_file_dialog() -> void:
 	_folder_creation_dialog.show()
 	return
+
 
 func _create_folder() -> void:
 	var _folder_name = _folder_creation_dialog.get_node("LineEdit").text
@@ -158,6 +191,7 @@ func _create_folder() -> void:
 	_folder_creation_dialog.get_node("LineEdit").text = ""
 	_build_view()
 	return
+
 
 func _open_delete_dialog() -> void:
 	if _selected_folder == null:
@@ -172,9 +206,25 @@ func _open_delete_dialog() -> void:
 	_deletion_dialog.show()
 	return
 
+
 func _delete_selected() -> void:
 	var _folder_name = _selected_folder.get_meta("label")
+	var _type: String = _selected_folder.get_meta("type")
 	GlobalLogger.log("Deleting Item '%s'" % _folder_name)
-	FileManager.delete_folder(_folder_name)
+
+	if _type == "spawnable":
+		var _file_name: String = _selected_folder.get_meta("file_name")
+		var _db_entry: Dictionary = Database.get_spawnable(_file_name)
+
+		if _db_entry.has("directory") == false:
+			GlobalLogger.log("Tried to delete file '%s' but the directory could not be found!" % _file_name, Enum.LogLevel.WARNING)
+			return
+
+		FileManager.delete_file(_db_entry.directory)
+		Database.delete_spawnable(_db_entry.hash)
+
+	if _type == "folder":
+		FileManager.delete_folder(_folder_name)
+
 	_build_view()
 	return
